--- id-logik-lookup/README.md
@@ -1,13 +1,0 @@
-# ID Logik Lookup Recommender
-
-Dieses Projekt stellt einen Recommender für INCEpTION in Form einer Spring Boot Application bereit.  
-Der Recommender bildet dann die Schnittstelle zwischen INCEpTION und dem ID Logik Server um Terme bzgl. SNOMED zu grounden.
-
-### Installation
-_ToDo_
-
-### Konfiguration
-_ToDo_
-
-### Anderes
-_ToDo_
