--- id-logik-lookup/src/main/java/com/id/inceptionrecommender/InceptionRecommenderApplication.java
@@ -1,13 +1,0 @@
-package com.id.inceptionrecommender;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class InceptionRecommenderApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(InceptionRecommenderApplication.class, args);
-    }
-
-}
