--- id-logik-lookup/src/test/java/com/id/inceptionrecommender/InceptionRecommenderApplicationTests.java
@@ -1,13 +1,0 @@
-package com.id.inceptionrecommender;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class InceptionRecommenderApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
