+++ id_logik_lookup/src/main/java/com/id/inceptionrecommender/rest/IdLogikServices.java
@@ -2,6 +2,7 @@ package com.id.inceptionrecommender.rest;
 
 import com.id.idlogik.app.ResultItem;
 import com.id.idlogik.app.Task;
+import com.id.shared.Sleep;
 import idlogik.IdLogikRestfulClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
@@ -20,27 +21,31 @@ public class IdLogikServices {
         idLogikRestfulClient = new IdLogikRestfulClient(config.getProtocol(), config.getHost(), config.getPort());
         idLogikRestfulClient.setLicence(config.getLicence());
         idLogikRestfulClient.login();
+        idLogikRestfulClient.setAutoLogin(true);
+        idLogikRestfulClient.setRetryAtNoSessionID(true);
+        idLogikRestfulClient.setMaxReconnect(10);
     }
 
-    @Cacheable(value = "snomedCTConcept", key = "#text")
-    public Optional<String> text2SingleSCT(String text) {
-        Task task = new Task();
-        task.setServiceName("term.Index");
-        task.getParameter().setProperty("SEARCHTEXT", text);
-        task.getParameter().setProperty("NOMENCLATURE_IDX", "1");
-        task.getParameter().setProperty("LANGUAGE", "0");
-        task.getParameter().setProperty("COMPOSITIONAL", "1");
-        task.getParameter().setProperty("USE_X_INDICES", "0");
-        task.getParameter().setProperty("EXTRACTION_PROFILE", "SEARCH");
-        idLogikRestfulClient.call(task);
-        if (task.getResultItems().size() == 1) {
-            return Optional.of(task.getResultItems().get(0).getAttributes().getProperty("NAME"));
-        } else {
-            return Optional.empty();
+    protected void call(Task task) {
+        call(task, 10);
+    }
+
+    /*
+     Diese Methode prüft, ob die Session-ID noch gültig ist und invalidiert diese ggf.
+     Es werden maximal 10 Versuche durchgeführt, wobei jeweils fünf Sekunden Wartezeit dazwischen liegt
+     (falls der ID LOGIK Server gerade startet)
+     */
+    protected void call(Task task, int retries) {
+        final int status = idLogikRestfulClient.call(task);
+        if (retries > 0 && status == 1008) {
+            task.getParameter().remove("SESSION_ID");
+            idLogikRestfulClient.setSessionID("");
+            Sleep.doSleep(5000);
+            call(task, retries-1);
         }
     }
 
-    @Cacheable(value = "text2FullSCT", key = "#text")
+    @Cacheable(value = "SCT_ANNO_CACHE", key = "#text", unless="#task.getResultItems().size()==0")
     public List<ResultItem> text2FullSCT(String text) {
         Task task = new Task();
         task.setServiceName("term.Index");
@@ -50,15 +55,8 @@ public class IdLogikServices {
         task.getParameter().setProperty("COMPOSITIONAL", "1");
         task.getParameter().setProperty("USE_X_INDICES", "0");
         task.getParameter().setProperty("EXTRACTION_PROFILE", "SEARCH");
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
-        /*
-        if (task.getResultItems().size() == 1) {
-            return Optional.of(task.getResultItems().get(0).getAttributes().getProperty("NAME"));
-        } else {
-            return Optional.empty();
-        }
-        */
     }
 
     public List<ResultItem> text2SCT(String text) {
@@ -70,7 +68,7 @@ public class IdLogikServices {
         task.getParameter().setProperty("COMPOSITIONAL", "1");
         task.getParameter().setProperty("USE_X_INDICES", "0");
         task.getParameter().setProperty("EXTRACTION_PROFILE", "SEARCH");
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
     }
 
@@ -82,7 +80,7 @@ public class IdLogikServices {
         task.getParameter().setProperty("CLASS_IDX", "265");
         task.getParameter().setProperty("CONTEXT", "0");
         task.getParameter().setProperty("MAX_RESULT_LINES", "1");
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
     }
 
@@ -93,7 +91,7 @@ public class IdLogikServices {
         task.getParameter().setProperty("YEAR", "2023");
         task.getParameter().setProperty("COUNTRY", "DE");
         task.getParameter().setProperty("FORMAT", "%I%C?%T");
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
     }
 
@@ -101,7 +99,7 @@ public class IdLogikServices {
         Task task = new Task();
         task.setServiceName("term.GetConcept");
         task.getParameter().setProperty("IDENT_NR", code);
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
     }
 
@@ -113,8 +111,25 @@ public class IdLogikServices {
         task.getParameter().setProperty("LANGUAGE", "0");
         task.getParameter().setProperty("TEXT_ONLY", "1");
         task.getParameter().setProperty("DETAIL_MODE", "0");
-        idLogikRestfulClient.call(task);
+        call(task);
         return task.getResultItems();
+    }
+
+    public Optional<String> text2SingleSCT(String text) {
+        Task task = new Task();
+        task.setServiceName("term.Index");
+        task.getParameter().setProperty("SEARCHTEXT", text);
+        task.getParameter().setProperty("NOMENCLATURE_IDX", "1");
+        task.getParameter().setProperty("LANGUAGE", "0");
+        task.getParameter().setProperty("COMPOSITIONAL", "1");
+        task.getParameter().setProperty("USE_X_INDICES", "0");
+        task.getParameter().setProperty("EXTRACTION_PROFILE", "SEARCH");
+        call(task);
+        if (task.getResultItems().size() == 1) {
+            return Optional.of(task.getResultItems().get(0).getAttributes().getProperty("NAME"));
+        } else {
+            return Optional.empty();
+        }
     }
 
 }
