+++ id_logik_lookup/src/main/resources/application.properties
